Fix for "Invalid property" exception when polyfilling an object with a getter
